Move grammar module to top-level in eth-abi
